requirements.txt
numpy
torch>=2.2
torchvision
einops
tqdm
pyyaml
pillow
pandas
pyarrow
tensorboardX
pytest
transformers==4.40.1
tokenizers==0.19.1
timm==0.9.10
accelerate
meshcat
imageio[ffmpeg]
manipulation
ipython
lxml
matplotlib
mpld3
pyvirtualdisplay
trimesh
triangle
scipy
torchcfm